+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - CASMCMS-9561: Consolidate back-end code for multi-session-delete endpoints
     - CASMCMS-9562: Consolidate back-end code for multi-component-patch endpoints
     - CASMCMS-9564: Create `DBNoEntryError` and `DBTooBusyError` exceptions
+    - CASMCMS-9554: Update DBWrapper `delete`/`get_delete` calls to use `DBNoEntryError`
 
 ## [1.27.0] - 07/03/2025
 ### Changed